refactor validation error formatting and fix cache condition
