+++ src/types.ts
@@ -1,3 +1,11 @@
+export interface ValidationError {
+  root: string;
+  path: string;
+  keyword: string;
+  schemaPath: string;
+  expected: string;
+}
+
 export interface BaseCertificateSchema {
   RefSchemaUrl: string;
   [key: string]: any;
@@ -21,3 +29,5 @@ export interface ECoCSchema extends BaseCertificateSchema {
   URL: string;
   EcocData: any;
 }
+
+export type SchemaTypes = 'en10168-schemas' | 'e-coc-schemas';
